Fix nav flicker by stable state hooks
